fix a line beark
