+++ app/build.gradle.kts
@@ -34,7 +34,7 @@ android {
         jvmTarget = "1.8"
     }
 
-    buildFeatures{
+    buildFeatures {
         viewBinding = true
     }
 }
@@ -54,4 +54,28 @@ dependencies {
     // navigation
     implementation("androidx.navigation:navigation-ui-ktx:2.7.7")
     implementation("androidx.navigation:navigation-fragment-ktx:2.7.7")
+    // Retrofit
+    implementation("com.squareup.retrofit2:retrofit:2.10.0")
+    implementation("com.squareup.retrofit2:converter-gson:2.10.0")
+    // Gson
+    implementation("com.google.code.gson:gson:2.10.1")
+    // Coroutine
+    implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.7.3")
+    // ViewModel
+    implementation("androidx.lifecycle:lifecycle-viewmodel-ktx:2.7.0")
+    // LiveData
+    implementation("androidx.lifecycle:lifecycle-livedata-ktx:2.7.0")
+    // Dagger2
+    implementation ("com.google.dagger:dagger:2.35.1")
+    implementation ("com.google.dagger:dagger-android:2.35.1")
+    implementation ("com.google.dagger:dagger-android-support:2.26")
+//    kapt ("com.google.dagger:dagger-compiler:2.31.2")
+//    kapt ("com.google.dagger:dagger-android-processor:2.26")
+
+    // okhttp3
+    implementation ("com.squareup.okhttp3:logging-interceptor:5.0.0-alpha.2")
+    // swipe refresh layout
+    implementation ("androidx.swiperefreshlayout:swiperefreshlayout:1.1.0")
+    // glide
+    implementation ("com.github.bumptech.glide:glide:4.16.0")
 }
