+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools">
+    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
+    <uses-permission android:name="android.permission.INTERNET" />
 
     <application
         android:allowBackup="true"
@@ -13,22 +15,19 @@
         android:theme="@style/Theme.CarRentalMVVM"
         tools:targetApi="31">
         <activity
-            android:name=".ui.SearchActivity"
+            android:name=".ui.getstarted.GetStartedActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.GetStartedActivity"
+            android:name=".ui.signin.SignInActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.SignInActivity"
-            android:exported="false" />
-        <activity
-            android:name=".ui.SignUpActivity"
+            android:name=".ui.signup.SignUpActivity"
             android:exported="false" />
         <activity
             android:name=".MainActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.SplashActivity"
+            android:name=".ui.splash.SplashActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
