+++ app/src/main/java/com/carrentalapp/mvvm/data/model/Cars.kt
@@ -1,0 +1,6 @@
+package com.carrentalapp.mvvm.data.model
+
+data class Cars(
+    val name : String,
+    val price : Int
+)
