+++ app/src/main/java/com/carrentalapp/mvvm/data/model/User.kt
@@ -1,0 +1,7 @@
+package com.carrentalapp.mvvm.data.model
+
+class User(
+    val id : Int,
+    val username : String,
+    val password : String
+)
