--- app/src/main/java/com/carrentalapp/mvvm/model/Cars.kt
@@ -1,6 +1,0 @@
-package com.carrentalapp.mvvm.model
-
-data class Cars (
-    val name: String,
-    val price: Int
-)
