+++ app/src/main/java/com/carrentalapp/mvvm/ui/favorite/FavoriteFragment.kt
@@ -1,0 +1,58 @@
+package com.carrentalapp.mvvm.ui.favorite
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.recyclerview.widget.LinearLayoutManager
+
+import com.carrentalapp.mvvm.adapter.CarsFavoritesAdapter
+import com.carrentalapp.mvvm.data.model.Cars
+import com.carrentalapp.mvvm.databinding.FragmentFavoriteBinding
+
+
+
+
+class FavoriteFragment : Fragment() {
+
+    private var _binding: FragmentFavoriteBinding? = null
+
+    private val binding get() = _binding!!
+
+    private val carsFavoriteAdapter = CarsFavoritesAdapter()
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        _binding = FragmentFavoriteBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        setupRecyclerView()
+
+
+        val cars = List(100) { index ->
+            Cars(
+                name = "TOYOTA", price = 300
+            )
+        }
+        carsFavoriteAdapter.setCars(cars)
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+
+    private fun setupRecyclerView() {
+        binding.recyclerView.apply {
+            setHasFixedSize(true)
+            layoutManager = LinearLayoutManager(context, LinearLayoutManager.VERTICAL, false)
+            adapter = carsFavoriteAdapter
+        }
+    }
+}
