--- app/src/main/java/com/carrentalapp/mvvm/ui/fragments/PersonFragment.kt
@@ -1,22 +1,0 @@
-package com.carrentalapp.mvvm.ui.fragments
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.carrentalapp.mvvm.R
-class PersonFragment : Fragment() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-    }
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_person, container, false)
-    }
-}
