+++ app/src/main/java/com/carrentalapp/mvvm/ui/home/HomeFragment.kt
@@ -1,0 +1,62 @@
+package com.carrentalapp.mvvm.ui.home
+
+
+import android.content.Intent
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.recyclerview.widget.GridLayoutManager
+import com.carrentalapp.mvvm.adapter.CarsAdapter
+import com.carrentalapp.mvvm.data.model.Cars
+import com.carrentalapp.mvvm.databinding.FragmentHomeBinding
+import com.carrentalapp.mvvm.ui.search.SearchActivity
+
+
+class HomeFragment : Fragment() {
+
+    private var _binding: FragmentHomeBinding? = null
+
+    private val binding get() = _binding!!
+
+    private val carsAdapter = CarsAdapter()
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        _binding = FragmentHomeBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        setupRecyclerView()
+
+        binding.btnSearch.setOnClickListener {
+            val intent = Intent(requireContext(), SearchActivity::class.java)
+            startActivity(intent)
+        }
+
+        val cars = List(100) { index ->
+            Cars(
+                name = "TOYOTA", price = 300
+            )
+        }
+        carsAdapter.setCars(cars)
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+
+    private fun setupRecyclerView() {
+        binding.recyclerView.apply {
+            setHasFixedSize(true)
+            layoutManager = GridLayoutManager(context, 2)
+            adapter = carsAdapter
+        }
+    }
+}
