+++ app/src/main/java/com/carrentalapp/mvvm/ui/splash/SplashActivity.kt
@@ -1,0 +1,29 @@
+package com.carrentalapp.mvvm.ui.splash
+
+import android.content.Intent
+import android.os.Bundle
+import android.os.Handler
+import android.view.WindowManager
+import androidx.appcompat.app.AppCompatActivity
+import com.carrentalapp.mvvm.databinding.ActivitySplashBinding
+import com.carrentalapp.mvvm.ui.getstarted.GetStartedActivity
+
+class SplashActivity : AppCompatActivity() {
+    private lateinit var binding: ActivitySplashBinding
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivitySplashBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        supportActionBar?.hide()
+        this.window.setFlags(
+            WindowManager.LayoutParams.FLAG_FULLSCREEN,
+            WindowManager.LayoutParams.FLAG_FULLSCREEN,
+        )
+
+        Handler().postDelayed({
+            startActivity(Intent(this, GetStartedActivity::class.java))
+            finish()
+        },3000)
+    }
+}
